+++ primefaces/src/main/java/org/primefaces/component/autocomplete/AutoCompleteRenderer.java
@@ -580,22 +580,20 @@ public class AutoCompleteRenderer extends InputRenderer {
         writer.startElement("tbody", ac);
 
         if (items != null) {
-            int rowNumber = 0;
+            int index = 0;
             if (ac.isClientQueryMode() || items instanceof Map) {
                 for (Map.Entry<String, List<String>> entry : ((Map<String, List<String>>) items).entrySet()) {
                     String key = entry.getKey();
                     List<String> list = entry.getValue();
 
                     for (Object item : list) {
-                        encodeSuggestionItemsAsTable(context, ac, item, converter, pojo, var, key, rowNumber);
-                        rowNumber++;
+                        encodeSuggestionItemsAsTable(context, ac, item, converter, pojo, var, key, index++);
                     }
                 }
             }
             else {
                 for (Object item : (List) items) {
-                    encodeSuggestionItemsAsTable(context, ac, item, converter, pojo, var, null, rowNumber);
-                    rowNumber++;
+                    encodeSuggestionItemsAsTable(context, ac, item, converter, pojo, var, null, index++);
                 }
 
                 if (ac.hasMoreSuggestions()) {
@@ -619,22 +617,20 @@ public class AutoCompleteRenderer extends InputRenderer {
         writer.writeAttribute(HTML.ARIA_ROLE, HTML.ARIA_ROLE_LISTBOX, null);
 
         if (items != null) {
-            int rowNumber = 0;
+            int index = 0;
             if (ac.isClientQueryMode() || items instanceof Map) {
                 for (Map.Entry<String, List<String>> entry : ((Map<String, List<String>>) items).entrySet()) {
                     String key = entry.getKey();
                     List<String> list = entry.getValue();
 
                     for (Object item : list) {
-                        encodeSuggestionItemsAsList(context, ac, item, converter, pojo, var, key, rowNumber);
-                        rowNumber++;
+                        encodeSuggestionItemsAsList(context, ac, item, converter, pojo, var, key, index++);
                     }
                 }
             }
             else {
                 for (Object item : (List) items) {
-                    encodeSuggestionItemsAsList(context, ac, item, converter, pojo, var, null, rowNumber);
-                    rowNumber++;
+                    encodeSuggestionItemsAsList(context, ac, item, converter, pojo, var, null, index++);
                 }
 
                 if (ac.hasMoreSuggestions()) {
@@ -664,8 +660,9 @@ public class AutoCompleteRenderer extends InputRenderer {
         if (pojo) {
             requestMap.put(var, item);
             String value = converter == null ? String.valueOf(ac.getItemValue()) : converter.getAsString(context, ac, ac.getItemValue());
+            String itemLabel = ac.getItemLabel();
             writer.writeAttribute("data-item-value", value, null);
-            writer.writeAttribute("data-item-label", ac.getItemLabel(), null);
+            writer.writeAttribute("data-item-label", itemLabel, null);
             writer.writeAttribute("data-item-class", ac.getItemStyleClass(), null);
             writer.writeAttribute("data-item-group", ac.getGroupBy(), null);
 
@@ -677,18 +674,29 @@ public class AutoCompleteRenderer extends InputRenderer {
                 writer.writeAttribute("data-item-group-tooltip", ac.getGroupByTooltip(), null);
             }
 
-            writer.writeText(ac.getItemLabel(), null);
+            if (ac.isEscape()) {
+                writer.writeText(itemLabel, null);
+            }
+            else {
+                writer.write(itemLabel);
+            }
         }
         else {
-            writer.writeAttribute("data-item-label", item.toString(), null);
-            writer.writeAttribute("data-item-value", item.toString(), null);
+            String itemAsString = item.toString();
+            writer.writeAttribute("data-item-label", itemAsString, null);
+            writer.writeAttribute("data-item-value", itemAsString, null);
             writer.writeAttribute("data-item-class", ac.getItemStyleClass(), null);
 
             if (key != null) {
                 writer.writeAttribute("data-item-key", key, null);
             }
 
-            writer.writeText(item, null);
+            if (ac.isEscape()) {
+                writer.writeText(item, null);
+            }
+            else {
+                writer.write(itemAsString);
+            }
         }
 
         writer.endElement("li");
@@ -702,14 +710,14 @@ public class AutoCompleteRenderer extends InputRenderer {
     }
 
     protected void encodeSuggestionItemsAsTable(FacesContext context, AutoComplete ac, Object item, Converter converter,
-            boolean pojo, String var, String key, int rowNumber) throws IOException {
+            boolean pojo, String var, String key, int index) throws IOException {
         ResponseWriter writer = context.getResponseWriter();
         Map<String, Object> requestMap = context.getExternalContext().getRequestMap();
         UIComponent itemtip = ac.getFacet("itemtip");
         boolean hasGroupByTooltip = (ac.getValueExpression(AutoComplete.PropertyKeys.groupByTooltip.toString()) != null);
 
         writer.startElement("tr", null);
-        writer.writeAttribute("id", ac.getClientId(context) + "_item_" + rowNumber, null);
+        writer.writeAttribute("id", ac.getClientId(context) + "_item_" + index, null);
         writer.writeAttribute("class", AutoComplete.ROW_CLASS, null);
 
         if (pojo) {
@@ -743,7 +751,7 @@ public class AutoCompleteRenderer extends InputRenderer {
                     writer.writeAttribute("class", column.getStyleClass(), null);
                 }
 
-                column.encodeAll(context);
+                encodeIndexedId(context, column, index);
 
                 writer.endElement("td");
             }
@@ -752,7 +760,7 @@ public class AutoCompleteRenderer extends InputRenderer {
         if (ComponentUtils.shouldRenderFacet(itemtip)) {
             writer.startElement("td", null);
             writer.writeAttribute("class", AutoComplete.ITEMTIP_CONTENT_CLASS, null);
-            itemtip.encodeAll(context);
+            encodeIndexedId(context, itemtip, index);
             writer.endElement("td");
         }