+++ primefaces/src/main/java/org/primefaces/component/badge/Badge.java
@@ -27,6 +27,7 @@ import org.primefaces.model.badge.BadgeModel;
 import org.primefaces.model.badge.DefaultBadgeModel;
 
 import javax.faces.application.ResourceDependency;
+import javax.faces.context.FacesContext;
 
 @ResourceDependency(library = "primefaces", name = "components.css")
 public class Badge extends BadgeBase {
@@ -42,6 +43,14 @@ public class Badge extends BadgeBase {
     public static final String SEVERITY_SUCCESS_CLASS = "ui-badge-success";
     public static final String SEVERITY_WARNING_CLASS = "ui-badge-warning";
     public static final String SEVERITY_DANGER_CLASS = "ui-badge-danger";
+
+    public BadgeRenderer getRenderer() {
+        return (BadgeRenderer) getFacesContext().getRenderKit().getRenderer(getFamily(), getRendererType());
+    }
+
+    public static Badge create(FacesContext context) {
+        return (Badge) context.getApplication().createComponent(Badge.COMPONENT_TYPE);
+    }
 
     public BadgeModel toBadgeModel() {
         return DefaultBadgeModel.builder()