+++ primefaces/src/main/java/org/primefaces/component/datatable/feature/DataTableFeatures.java
@@ -24,6 +24,7 @@
 package org.primefaces.component.datatable.feature;
 
 import java.util.Collection;
+import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
@@ -32,6 +33,7 @@ import java.util.stream.Stream;
 
 public class DataTableFeatures {
 
+    // order matters: filter should be decoded before selection
     private static final Map<Class<? extends DataTableFeature>, DataTableFeature> ALL_FEATURES = Stream.of(
                     new DraggableColumnsFeature(),
                     new FilterFeature(),
@@ -45,7 +47,7 @@ public class DataTableFeatures {
                     new ScrollFeature(),
                     new DraggableRowsFeature(),
                     new AddRowFeature())
-            .collect(Collectors.toMap(DataTableFeature::getClass, Function.identity()));
+            .collect(Collectors.toMap(DataTableFeature::getClass, Function.identity(), (u, v) -> u, LinkedHashMap::new));
 
     private DataTableFeatures() {
         // NOOP