+++ primefaces/src/main/resources/META-INF/resources/primefaces/overlaypanel/overlaypanel.js
@@ -60,8 +60,11 @@ PrimeFaces.widget.OverlayPanel = PrimeFaces.widget.DynamicOverlayWidget.extend({
      * @param {PrimeFaces.PartialWidgetCfg<TCfg>} cfg
      */
     init: function(cfg) {
-       if (cfg.target) {
+        if (cfg.target) {
             this.target = PrimeFaces.expressions.SearchExpressionFacade.resolveComponentsAsSelector(cfg.target);
+            if (this.target.hasClass('ui-splitbutton')) {
+                this.target = this.target.find('.ui-splitbutton-menubutton');
+            }
         }
         this._super(cfg, null, null, this.target);
 
@@ -402,6 +405,9 @@ PrimeFaces.widget.OverlayPanel = PrimeFaces.widget.DynamicOverlayWidget.extend({
             allowedNegativeValuesByParentOffset = this.jq.offsetParent().offset();
 
         this.targetElement = this.getTarget(target);
+        if (this.targetElement.hasClass('ui-splitbutton-menubutton')) {
+            this.targetElement = this.targetElement.parent();
+        }
         if (this.targetElement) {
             this.targetZindex = this.targetElement.zIndex();
         }